Confused on the laundry request form
I found that if I enter the date in a wrong format, i.e., 01212021, nothing would happen but the text tells me "Loading". It may let users wait for a long time but no further changes, which will lead to bad user experience. Could you please add a format checker like the one you used in sign up form? The page is not responding with anything even though I entered the date as 20/12/2021. Is it because there is no data in your database?